2023/Web/beer_me_up_before_you_format/requirements.txt
flask==3.0.0
pyjwt==2.8.0